+++ CHANGES.md
@@ -1,5 +1,13 @@
 # NGI Python KOF Parser Package
 
+## Version 0.0.7
+_2022-04-20_
+
+Add
+
+- Transformation of coordinates in output file.
+- Add support for swapping easting and northing in the input or output kof file.
+
 ## Version 0.0.6
 _2022-03-02_